requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
aiofiles
python-dotenv
asyncpg
databases